refactor blogsRouter

+++ part4/bloglist/controllers/blogs.js
@@ -25,16 +25,8 @@ blogsRouter.get('/:id', (request, response) => {
 })
 
 blogsRouter.post('/', userExtractor, async (request, response) => {
-  const body = request.body
   const user = request.user
-  const blog = new Blog({
-    likes: body.likes,
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    user: user.id,
-  })
-
+  const blog = new Blog({ ...request.body, user: user.id })
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -44,22 +36,25 @@ blogsRouter.post('/', userExtractor, async (request, response) => {
 blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const user = request.user
   const blogToDelete = await Blog.findById(request.params.id)
-
-  if (blogToDelete.user.toString() !== user.id.toString()) {
-    return response.status(401).json({ error: 'unauthorized to delete ' })
+  if (!blogToDelete) {
+    return response.status(204).end()
   }
-
+  if (blogToDelete.user.toString() !== user.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: 'only the creator can delete a blog' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
-  const blog = {
-    likes: request.body.likes,
-  }
+  const blog = request.body
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
     new: true,
+    runValidators: true,
+    context: 'query',
   })
 
   response.json(updatedBlog)
